requirements.txt
streamlit
torch
torchvision
opencv-python
numpy
Pillow
ultralytics
google-generativeai
python-dotenv